updates data type to uint32 for saving

+++ eosdxanalysis/preprocessing/preprocess.py
@@ -646,13 +646,13 @@ class PreprocessData(object):
                 if output_format == "txt":
                     save_filename = "preprocessed_{}".format(filename)
                     save_filename_fullpath = os.path.join(output_dir, save_filename)
-                    np.savetxt(save_filename_fullpath,output.astype(np.uint16),fmt='%i')
+                    np.savetxt(save_filename_fullpath,output.astype(np.uint32),fmt='%i')
 
                 # Save output as image
                 if output_format == "png":
                     save_filename = "preprocessed_{}.png".format(filename)
                     save_filename_fullpath = os.path.join(output_dir, save_filename)
-                    imageio.imwrite(save_filename_fullpath, output.astype(np.uint16))
+                    imageio.imwrite(save_filename_fullpath, output.astype(np.uint32))
 
 
 if __name__ == "__main__":
